refactor:#6 데이터 가공 로직 분리
